fix bug (infinite recursion)

+++ app/src/main/java/jsh/utils/ValidateUtil.java
@@ -20,27 +20,27 @@ public final class ValidateUtil {
     // methods for other primitive types
 
     public static int requireNaturalNumber(final int n, final String message) {
-        return requireNaturalNumber(n, message);
+        return (int) requireNaturalNumber((long) n, message);
     }
 
     public static short requireNaturalNumber(final short n, final String message) {
-        return requireNaturalNumber(n, message);
+        return (short) requireNaturalNumber((long) n, message);
     }
 
     public static byte requireNaturalNumber(final byte n, final String message) {
-        return requireNaturalNumber(n, message);
+        return (byte) requireNaturalNumber((long) n, message);
     }
 
     public static int requireNaturalNumber(final int n) {
-        return requireNaturalNumber(n);
+        return (int) requireNaturalNumber((long) n);
     }
 
     public static short requireNaturalNumber(final short n) {
-        return requireNaturalNumber(n);
+        return (short) requireNaturalNumber((long) n);
     }
 
     public static byte requireNaturalNumber(final byte n) {
-        return requireNaturalNumber(n);
+        return (byte) requireNaturalNumber((long) n);
     }
 
     private static boolean isNaturalNumber(final long n) {
